+++ Changelog.md
@@ -91,7 +91,7 @@ Lilu Changelog
 #### v1.4.8
 - Added MacKernelSDK with Xcode 12 compatibility
 - Removed `kern_atomic.hpp` due to MacKernelSDK implementation
-- Acidanthera MacKernelSDK is now required for all plugins
+- devicemanager MacKernelSDK is now required for all plugins
 - Fixed Lilu loading on macOS 10.6 (not all APIs will be functional)
 - Fixed plugin debug log not working with Lilu disabled